pyproject.toml
[project]
name = "schema-migration-openenv"
version = "1.0.0"
description = "A real-world RL environment for database schema migrations"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "pydantic==2.7.1",
    "openai>=2.7.2",
    "requests==2.32.2",
    "python-multipart==0.0.9",
]

[project.scripts]
server = "server.app:main"
serve = "server.app:main"
schema-migration-openenv = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=68", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "tasks*", "graders*", "server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
